move to scanner_uilts/model.py
